+++ src/openvdb/utils/primitives/MeshVolume.js
@@ -1,4 +1,5 @@
 import * as Three from 'three';
+import * as BufferGeometryUtils from 'three/examples/jsm/utils/BufferGeometryUtils';
 import { floatType } from '../../math/memory';
 import { GridDescriptor } from '../../core/GridDescriptor';
 import { Vector3 } from '../../math/vector';
@@ -22,13 +23,51 @@ export class MeshVolume extends GridDescriptor {
 
     // NOTE Primitives don't need to use OpenVDBReader - just "read grid" right away
     this.readGrid();
-    this.mesh = mesh.clone();
+    
+    const meshClone = mesh.clone();
+    const geometries = [];
+    const transform = new Three.Vector3();
+    const transformQuaternion = new Three.Quaternion();
 
-    this.mesh.traverse((child) => {
-      if (child.material) {
-        child.material = new Three.MeshBasicMaterial({ color: 0xff00ff, side: Three.DoubleSide });
+    meshClone.position.setScalar(0.0, 0.0, 0.0);
+    meshClone.scale.setScalar(1.0);
+    meshClone.quaternion.identity();
+
+    meshClone.updateMatrix();
+    meshClone.updateMatrixWorld();
+    meshClone.updateWorldMatrix();
+
+    meshClone.traverse((child) => {
+      if (child.geometry) {
+        const geometry = child.geometry.clone();
+
+        child.getWorldQuaternion(transformQuaternion);
+        geometry.applyQuaternion(transformQuaternion);
+
+        child.getWorldScale(transform);
+        geometry.scale(transform.x, transform.y, transform.z);
+
+        child.getWorldPosition(transform);
+        geometry.translate(transform.x, transform.y, transform.z);
+
+        geometry.attributes = {
+          position: geometry.attributes.position,
+          normal: geometry.attributes.normal,
+        };
+
+        geometries.push(geometry);
       }
     });
+
+    // NOTE Merge all geometries into a single mesh to avoid raycasting tons of separate objects
+    const meshGeometry = BufferGeometryUtils.mergeVertices(
+      BufferGeometryUtils.mergeBufferGeometries(geometries, false)
+    );
+
+    this.mesh = new Three.Mesh(
+      meshGeometry,
+      new Three.MeshNormalMaterial({ side: Three.DoubleSide })
+    );
 
     this.cacheMeshGeometry();
   }
@@ -97,7 +136,15 @@ export class MeshVolume extends GridDescriptor {
 
     this.bbox.getCenter(this.center);
     this.bbox.getSize(this.size);
-    this.size.addScalar(0.1); // NOTE Add padding to avoid edge glitches
+
+    const maxModelSpan = Math.max(this.size.x, this.size.y, this.size.z);
+    const sizeNormalization = new Three.Vector3().copy(this.size).divideScalar(maxModelSpan);
+
+    // NOTE Models can have various real aspect ratios - normalize the volumetric sampler to avoid model stretching
+    this.size.divide(sizeNormalization);
+
+    // NOTE Add padding to avoid edge glitches
+    this.size.addScalar(0.1);
 
     this.raycaster.near = -0.01;
     this.raycaster.far = 1.0; // NOTE We can assume this since the output volume is always 1x1x1
@@ -108,7 +155,7 @@ export class MeshVolume extends GridDescriptor {
     this.rayDirection.copy(position).normalize();
     this.raycaster.set(this.target, this.rayDirection);
 
-    const hits = this.raycaster.intersectObject(this.mesh, true);
+    const hits = this.raycaster.intersectObjects([ this.mesh ], true);
     
     return (hits.length === 0 || hits.length % 2 === 0) ? 0.0 : 1.0;
   }